requirements.txt
accelerate==0.18.0
anthropic==0.2.7
arrow==1.2.3
beautifulsoup4==4.12.0
bitsandbytes==0.38.1
datasets==2.7.1
detectron2==0.5
editdistance==0.6.2
evaluate==0.4.0
faker==18.4.0
fire==0.5.0
flask==2.2.3
flask-babel==3.1.0
future==0.18.3
huggingface-hub==0.13.3
jinja2==3.1.2
matplotlib==3.7.1
numpy==1.24.2
openai==0.27.2
opencv-python==4.6.0.66
pandas==1.5.3
pillow==9.5.0
pyarrow==11.0.0
requests==2.28.2
rouge-score==0.1.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.0.dev0
wandb==0.15.0
